index.htm
A website to control an LED Matrix. Just put the file on a mircro sdcard, plug it into your Arduino ethernet shield. The provided arduino sketch will send this website to your browser.
You will see a grid and a color picker. Pick a color and click on the grid to light up an LED.